+++ lascli/parser/users.py
@@ -6,10 +6,6 @@ from las import Client
 from lascli.util import nullable, NotProvided
 
 
-def encode_avatar(avatar):
-    return base64.b64encode(pathlib.Path(avatar).read_bytes()).decode()
-
-
 def list_users(las_client: Client, **optional_args):
     return las_client.list_users(**optional_args)
 
@@ -18,20 +14,29 @@ def get_user(las_client: Client, user_id):
     return las_client.get_user(user_id)
 
 
-def create_user(las_client: Client, email, app_client_id, **optional_args):
-    avatar = optional_args.get('avatar')
-    if avatar:
-        optional_args['avatar'] = encode_avatar(avatar)
+def create_user(las_client: Client, email, app_client_id, role_ids, **optional_args):
+    args = dict(
+        email=email,
+        app_client_id=app_client_id,
+        **optional_args,
+    )
+    
+    if role_ids:
+        args['role_ids'] = role_ids
+        
+    return las_client.create_user(**args)
 
-    return las_client.create_user(email, app_client_id=app_client_id, **optional_args)
 
+def update_user(las_client: Client, user_id, role_ids, **optional_args):
+    args = dict(
+        user_id=user_id,
+        **optional_args,
+    )
 
-def update_user(las_client: Client, user_id, **optional_args):
-    avatar = optional_args.get('avatar')
-    if avatar:
-        optional_args['avatar'] = encode_avatar(avatar)
+    if role_ids:
+        args['role_ids'] = role_ids
 
-    return las_client.update_user(user_id, **optional_args)
+    return las_client.update_user(**args)
 
 
 def delete_user(las_client: Client, user_id):
@@ -45,20 +50,11 @@ def create_users_parser(subparsers):
     create_user_parser = subparsers.add_parser('create')
     create_user_parser.add_argument('email')
     create_user_parser.add_argument('app_client_id')
-    create_user_parser.add_argument('--name')
-    create_user_parser.add_argument('--avatar', help='Path to avatar JPEG image.')
     create_user_parser.add_argument('--role-ids', nargs='+')
     create_user_parser.set_defaults(cmd=create_user)
 
     update_user_parser = subparsers.add_parser('update')
     update_user_parser.add_argument('user_id')
-    update_user_parser.add_argument('--name', type=nullable(str), default=NotProvided)
-    update_user_parser.add_argument(
-        '--avatar',
-        help='Path to avatar JPEG image or "null" to remove avatar from user.',
-        type=nullable(str),
-        default=NotProvided
-    )
     update_user_parser.add_argument('--role-ids', nargs='+')
     update_user_parser.set_defaults(cmd=update_user)