+++ tests/service.py
@@ -79,8 +79,12 @@ def create_role_id():
     return f'las:role:{uuid4().hex}'
 
 
+def create_email():
+    return 'foo@bar.com'
+
+
 def create_error_config():
-    return {'email': 'foo@bar.com'}
+    return {'email': create_email()}
 
 
 def create_completed_config():